Do not create multiple `ResolvedArtifactSet` instances for a given transformed variant of a project variant.
There were several ways multiple instances could be created. Firstly, an instance was created when the task dependencies of a `Configuration` were visited and another later when the full graph was resolved. Secondly, this was done for each configuration or view that contained the output variant.

This change reduces heap allocation but also allows the instance to track some state about the inputs and result of the transformation.